+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export { Header } from "./Header";
+export { ItemList } from "./ItemList";
+export { ComplexForm } from "./ComplexForm";
+export { NotificationSystem } from "./NotificationSystem";
