+++ src/context/ThemeContext.tsx
@@ -1,0 +1,40 @@
+import { createContext, PropsWithChildren, useContext, useState } from "react";
+import { useCallback, useMemo } from "../@lib";
+
+interface ThemeContextType {
+  theme: string;
+  toggleTheme: () => void;
+}
+
+const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
+
+export const ThemeProvider: React.FC<PropsWithChildren> = ({ children }) => {
+  const [theme, setTheme] = useState("light");
+
+  const toggleTheme = useCallback(() => {
+    setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
+  }, []);
+
+  const contextValue: ThemeContextType = useMemo(
+    () => ({
+      theme,
+      toggleTheme,
+    }),
+    [theme, toggleTheme],
+  );
+
+  return (
+    <ThemeContext.Provider value={contextValue}>
+      {children}
+    </ThemeContext.Provider>
+  );
+};
+
+export function useThemeContext() {
+  const context = useContext(ThemeContext);
+
+  if (context === undefined) {
+    throw new Error("useThemeContext must be used within an ThemeProvider");
+  }
+  return context;
+}
